+++ app/src/components/MeetingHeader/fragments/RoomTitle.tsx
@@ -1,14 +1,17 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { Typography, styled, Tooltip, Stack } from '@mui/material';
+import { Typography, styled, Tooltip, Stack, ThemeProvider } from '@mui/material';
+import { InfoButton } from '@opentalk/common';
 import { truncate } from 'lodash';
-import { useMemo } from 'react';
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import { createOpenTalkTheme } from '../../../assets/themes/opentalk';
 import { useAppSelector } from '../../../hooks';
 import { selectCurrentBreakoutRoom } from '../../../store/slices/breakoutSlice';
-import { selectEventInfo } from '../../../store/slices/roomSlice';
+import { selectEventInfo, selectRoomInfo } from '../../../store/slices/roomSlice';
+import MeetingDetailsDialog from './MeetingDetailsDialog';
 
 export const ROOM_TITLE_MAX_LENGTH = 17;
 
@@ -35,8 +38,10 @@ const RoomTitle = () => {
   const { t } = useTranslation();
   const currentBreakoutRoom = useAppSelector(selectCurrentBreakoutRoom);
   const eventInfo = useAppSelector(selectEventInfo);
+  const roomInfo = useAppSelector(selectRoomInfo);
+  const [meetingDetailsDialogOpen, setMeetingDetailsDialogOpen] = useState(false);
 
-  const title = useMemo(() => {
+  const getTitle = () => {
     if (currentBreakoutRoom) {
       return currentBreakoutRoom.name;
     }
@@ -46,17 +51,33 @@ const RoomTitle = () => {
     }
 
     return t('fallback-room-title');
-  }, [currentBreakoutRoom, eventInfo]);
+  };
 
-  const truncatedTitle = truncate(title, { length: ROOM_TITLE_MAX_LENGTH });
+  const truncatedTitle = truncate(getTitle(), { length: ROOM_TITLE_MAX_LENGTH });
 
   return (
-    <Container>
-      <Tooltip translate="no" title={title} describeChild>
+    <Container direction="row">
+      <Tooltip translate="no" title={getTitle()} describeChild>
         <RoomTitleTypograhy noWrap translate="no" variant="h1">
           {truncatedTitle}
         </RoomTitleTypograhy>
       </Tooltip>
+      {roomInfo && eventInfo?.meetingDetails && (
+        <>
+          <InfoButton
+            aria-label={t('room-title-info-button-aria-label')}
+            onClick={() => setMeetingDetailsDialogOpen(true)}
+          />
+          <ThemeProvider theme={createOpenTalkTheme()}>
+            <MeetingDetailsDialog
+              eventInfo={eventInfo}
+              roomInfo={roomInfo}
+              open={meetingDetailsDialogOpen}
+              onClose={() => setMeetingDetailsDialogOpen(false)}
+            />
+          </ThemeProvider>
+        </>
+      )}
     </Container>
   );
 };