+++ packages/common/components/index.ts
@@ -3,14 +3,15 @@
 // SPDX-License-Identifier: EUPL-1.2
 import AccordionItem from './AccordionItem';
 import CommonFormItem from './CommonFormItem';
+import { CopyTextField } from './CopyTextField';
 import { DurationField, DurationValueOptions } from './DurationField';
 import { ErrorFormMessage } from './ErrorFormMessage';
 import { FormWrapper, FormProps } from './FormWrapper';
-import { IconButton, AdornmentIconButton, CircularIconButton } from './IconButtons';
+import { IconButton, AdornmentIconButton, CircularIconButton, InfoButton } from './IconButtons';
 import { ParticipantAvatar, setLibravatarOptions } from './ParticipantAvatar';
 import ProgressBar from './ProgressBar';
 import SortPopoverMenu from './SortPopoverMenu';
-import TextField from './TextField';
+import { TextField } from './TextField';
 import Toggle from './Toggle';
 import ToolbarMenuUtils from './ToolbarMenuUtils';
 import VisuallyHiddenTitle from './VisuallyHiddenTitle';
@@ -40,10 +41,12 @@ export {
   IconButton,
   AdornmentIconButton,
   CircularIconButton,
+  InfoButton,
   SortPopoverMenu,
   Toggle,
   VisuallyHiddenTitle,
   ErrorFormMessage,
+  CopyTextField,
 };
 
 export type { DurationValueOptions, FormProps };